feat(base-estimator): add base estimator
