+++ src/estimating/__init__.py
@@ -1,4 +1,5 @@
 from .contract import Estimator  # isort: skip (contract must be first)
+from .base_estimator import BaseEstimator  # isort: skip (contract must be first)
 
 from .factory import EstimatorFactory
 from .linear_estimator import LinearEstimator
